refactor: Updated sign-in and sign-up routes
