+++ .env.test
@@ -1,13 +1,13 @@
-NODE_ENV="development"
+NODE_ENV="test"
 NEXT_PUBLIC_SITE_URL="http://localhost:3000"
 
 # AUTH with CLERK
 NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY={required}
 CLERK_SECRET_KEY={required}
-NEXT_PUBLIC_CLERK_SIGN_IN_URL="/sign-in"
-NEXT_PUBLIC_CLERK_SIGN_UP_URL="/sign-up"
-NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL="/"
-NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL="/"
+NEXT_PUBLIC_CLERK_SIGN_IN_URL="/login"
+NEXT_PUBLIC_CLERK_SIGN_UP_URL="/register"
+NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL="/channels/@me"
+NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL="/login"
 
 # This was inserted by `prisma init`:
 # Environment variables declared in this file are automatically made available to Prisma.
