+++ src/app/(auth)/(routes)/register/[[...register]]/page.tsx
