+++ src/data.ts
@@ -1,9 +1,10 @@
 import { nitroBannerType, nitroFAQCategoryType, nitroFAQType, nitroPlanFeatureProps, nitroPerkType, nitroPlanProps, NavigationMenuItemProps, socialLinkType, FooterMenuSectionType } from "@/types"
 
 /* Global */
-export const signInUrl = process.env.NEXT_PUBLIC_CLERK_SIGN_IN_URL ?? "/sign-in"
-export const signedInUrl = "/channels/@me"
-export const signUpUrl = process.env.NEXT_PUBLIC_CLERK_SIGN_UP_URL ?? "/sign-up"
+export const signInUrl = process.env.NEXT_PUBLIC_CLERK_SIGN_IN_URL ?? "/login"
+export const afterSignInUrl = process.env.NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL ?? "/channels/@me"
+export const signUpUrl = process.env.NEXT_PUBLIC_CLERK_SIGN_UP_URL ?? "/register"
+export const afterSignUpUrl = process.env.NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL ?? "/login"
 export const placeholderImageUrl = "/images/placeholder.svg"
 export const days = Array.from({ length: 31 }, (_, i) => ({
     value: String(i + 1),
@@ -506,14 +507,14 @@ export const nitroFAQList: nitroFAQType[] = [
     { categoryId: 1, question: "What’s the difference between Nitro and Nitro Basic?", answer: "Nitro is our most popular plan that unlocks access to all available Nitro perks, and Nitro Basic includes a selection of the most-loved Nitro perks that help you better express yourself. See the comparison chart above for a full breakdown!" },
     { categoryId: 1, question: "How much does Nitro cost?", answer: "Cost varies depending on your country and plan. You can check Nitro prices under User Settings &gt; Nitro in either the desktop or mobile apps." },
     { categoryId: 1, question: "Where can I buy Nitro?", answer: "You can purchase Nitro by visiting the Nitro tab in User Settings, or in the Nitro page under Home on desktop. On mobile, visit User Settings by tapping your user icon in the bottom bar, then select Get Nitro." },
-    { categoryId: 2, question: "What payment methods do you accept?", answer: "We accept a variety of payment methods, which you can learn more info on <a href=\"https://support.discord.com/hc/articles/360017693772\" class=\"link-16px\" tabindex=\"0\">here</a>!" },
-    { categoryId: 2, question: "Do you offer Nitro and Nitro Basic in localized pricing?", answer: "We are constantly working to expand the availability of localized pricing for Nitro and Nitro Basic. Check <a href=\"https://support.discord.com/hc/articles/4407269525911\" class=\"link-16px\" tabindex=\"0\">this article</a> for information on where localized pricing is currently available." },
-    { categoryId: 3, question: "Are Nitro gifts real?", answer: "Yes! You can purchase a Nitro gift through the Nitro page on desktop or the Nitro Gifting tab within mobile under User Settings. Additionally, you can send a quick gift in a channel by selecting the gift icon on desktop or mobile. If someone sends you a gift, you can see what it looks like <a href=\"https://support.discord.com/hc/articles/360020877112\" class=\"link-16px\" tabindex=\"0\">here</a>.&nbsp;You can also buy Discord Nitro directly on Amazon.com (United States only)." },
-    { categoryId: 3, question: "How can I buy a Discord Nitro gift card on Amazon.com?", answer: "Nitro gift cards on Amazon are a great way to gift Nitro to a friend, or have someone gift Nitro to you. Just click <a href=\"https://www.amazon.com/dp/B0CF4J58QK?maas=maas_adg_DE3B8B6262C4DC306922DC26C0CE6A2D_afap_abs&amp;ref_=aa_maas&amp;tag=maas&amp;th=1\" tabindex=\"0\"><span class=\"text-span\">this link</span></a> for a variety of Discord gift options, from 1 to 12-month Nitro subscriptions. Learn more about Discord gift cards on Amazon in this <a href=\"https://support.discord.com/hc/en-us/articles/360020877112\" class=\"link-407\" tabindex=\"0\">Help Center article</a>." },
+    { categoryId: 2, question: "What payment methods do you accept?", answer: `We accept a variety of payment methods, which you can learn more info on <a href="https://support.discord.com/hc/articles/360017693772" class="link-16px" tabindex="0">here</a>!` },
+    { categoryId: 2, question: "Do you offer Nitro and Nitro Basic in localized pricing?", answer: `We are constantly working to expand the availability of localized pricing for Nitro and Nitro Basic. Check <a href="https://support.discord.com/hc/articles/4407269525911" class="link-16px" tabindex="0">this article</a> for information on where localized pricing is currently available.` },
+    { categoryId: 3, question: "Are Nitro gifts real?", answer: `Yes! You can purchase a Nitro gift through the Nitro page on desktop or the Nitro Gifting tab within mobile under User Settings. Additionally, you can send a quick gift in a channel by selecting the gift icon on desktop or mobile. If someone sends you a gift, you can see what it looks like <a href="https://support.discord.com/hc/articles/360020877112" class="link-16px" tabindex="0">here</a>.&nbsp;You can also buy Discord Nitro directly on Amazon.com (United States only).` },
+    { categoryId: 3, question: "How can I buy a Discord Nitro gift card on Amazon.com?", answer: `Nitro gift cards on Amazon are a great way to gift Nitro to a friend, or have someone gift Nitro to you. Just click <a href="https://www.amazon.com/dp/B0CF4J58QK?maas=maas_adg_DE3B8B6262C4DC306922DC26C0CE6A2D_afap_abs&amp;ref_=aa_maas&amp;tag=maas&amp;th=1" tabindex="0"><span class="text-span">this link</span></a> for a variety of Discord gift options, from 1 to 12-month Nitro subscriptions. Learn more about Discord gift cards on Amazon in this <a href="https://support.discord.com/hc/en-us/articles/360020877112" class="link-407" tabindex="0">Help Center article</a>.` },
     { categoryId: 3, question: "Where to redeem Nitro codes?", answer: "You can redeem your Nitro code on the desktop and web apps through the Gift Inventory page in User Settings." },
-    { categoryId: 3, question: "Can I get free Nitro from a bot or a giveaway?", answer: "The only official way to receive Nitro is through a <a href=\"https://support.discord.com/hc/articles/360020877112\" class=\"link-16px\" tabindex=\"0\">gift</a> sent to you in chat, or by receiving a code that can be redeemed under Gift Inventory in User Settings on the desktop and web apps. We strongly encourage not clicking on any links from people you don’t know well. For more tips on staying safe on Discord, visit our <a href=\"https://discord.com/safetycenter\" class=\"link-16px\" tabindex=\"0\">Safety</a> page." },
-    { categoryId: 3, question: "Does Discord offer Nitro Promotions?", answer: "We are always working on ways to give you more with Nitro! Keep an eye out in the app, on our <a href=\"https://twitter.com/discord\" class=\"link-16px\" tabindex=\"0\">Twitter</a> , <a href=\"https://www.instagram.com/discord\" class=\"link-16px\" tabindex=\"0\">Instagram</a>, or <a href=\"https://discord.com/blog\" class=\"link-16px\" tabindex=\"0\">blog</a> to see when new promotions go live. :)" },
-    { categoryId: 4, question: "My Nitro subscription comes with Boosts. What are those?", answer: "Server Boosts are a way to show support for your favorite communities. Each Boost that you give to a server helps collectively unlock perks for everyone in the server to enjoy. Learn more about Server Boosting <a href=\"https://support.discord.com/hc/articles/360028038352\" class=\"link-16px\" tabindex=\"0\">here</a>!" },
+    { categoryId: 3, question: "Can I get free Nitro from a bot or a giveaway?", answer: `The only official way to receive Nitro is through a <a href="https://support.discord.com/hc/articles/360020877112" class="link-16px" tabindex="0">gift</a> sent to you in chat, or by receiving a code that can be redeemed under Gift Inventory in User Settings on the desktop and web apps. We strongly encourage not clicking on any links from people you don’t know well. For more tips on staying safe on Discord, visit our <a href="https://discord.com/safetycenter" class="link-16px" tabindex="0">Safety</a> page.` },
+    { categoryId: 3, question: "Does Discord offer Nitro Promotions?", answer: `We are always working on ways to give you more with Nitro! Keep an eye out in the app, on our <a href="https://twitter.com/discord" class="link-16px" tabindex="0">Twitter</a> , <a href="https://www.instagram.com/discord" class="link-16px" tabindex="0">Instagram</a>, or <a href="https://discord.com/blog" class="link-16px" tabindex="0">blog</a> to see when new promotions go live. :)` },
+    { categoryId: 4, question: "My Nitro subscription comes with Boosts. What are those?", answer: `Server Boosts are a way to show support for your favorite communities. Each Boost that you give to a server helps collectively unlock perks for everyone in the server to enjoy. Learn more about Server Boosting <a href="https://support.discord.com/hc/articles/360028038352" class="link-16px" tabindex="0">here</a>!` },
     { categoryId: 4, question: "Are Boosts permanent?", answer: "A Nitro subscription only comes with 2 Boosts that you can share to any server you belong to. The Boosts do not stack over time, and they are not permanent. You’ll get to keep the Boosts for as long as your Nitro subscription is active." },
-    { categoryId: 4, question: "Why is my Nitro not working?", answer: "If you’re having any trouble with your subscription, or have any questions, you can <a href=\"https://dis.gd/billing\" class=\"link-16px\" tabindex=\"0\">reach out</a> to our team and we’ll be happy to help!" },
+    { categoryId: 4, question: "Why is my Nitro not working?", answer: `If you’re having any trouble with your subscription, or have any questions, you can <a href="https://dis.gd/billing" class="link-16px" tabindex="0">reach out</a> to our team and we’ll be happy to help!` },
 ]
